+++ src/audio/module_adapter/CMakeLists.txt
@@ -10,8 +10,10 @@ if((NOT CONFIG_LIBRARY) OR CONFIG_LIBRARY_STATIC)
 	add_local_sources(sof
 		module/volume/volume_generic.c
 		module/volume/volume_hifi3.c
+		module/volume/volume_hifi4.c
 		module/volume/volume_generic_with_peakvol.c
 		module/volume/volume_hifi3_with_peakvol.c
+		module/volume/volume_hifi4_with_peakvol.c
 		module/volume/volume.c)
 	endif()
