+++ src/audio/module_adapter/module/volume/volume_hifi4_with_peakvol.c
@@ -1,13 +1,13 @@
 // SPDX-License-Identifier: BSD-3-Clause
 //
-// Copyright(c) 2018 Intel Corporation. All rights reserved.
+// Copyright(c) 2022 Intel Corporation. All rights reserved.
 //
-// Author: Tomasz Lauda <tomasz.lauda@linux.intel.com>
+// Author: Andrula Song <andrula.song@intel.com>
 
 /**
  * \file
- * \brief Volume HiFi3 processing implementation without peak volume detection
- * \authors Tomasz Lauda <tomasz.lauda@linux.intel.com>
+ * \brief Volume HiFi3 & 4 processing implementation with peak volume detection
+ * \authors Andrula Song <andrula.song@intel.com>
  */
 
 #include <sof/audio/buffer.h>
@@ -21,19 +21,12 @@ LOG_MODULE_DECLARE(volume_hifi3, CONFIG_SOF_LOG_LEVEL);
 
 #include <sof/audio/volume.h>
 
-#ifdef VOLUME_HIFI3
+#ifdef VOLUME_HIFI4
 
-#if (!CONFIG_COMP_PEAK_VOL)
+#if CONFIG_COMP_PEAK_VOL
+#include <xtensa/tie/xt_hifi4.h>
 
-#include <xtensa/tie/xt_hifi3.h>
-
-/**
- * \brief store volume gain 4 times for xtensa multi-way intrinsic operations.
- * Simultaneous processing 2 data.
- * \param[in,out] cd Volume component private data.
- * \param[in] channels_count Number of channels to process.
- */
-static void vol_store_gain(struct vol_data *cd, const int channels_count)
+static inline void vol_store_gain(struct vol_data *cd, const int channels_count)
 {
 	int32_t i;
 
@@ -49,12 +42,12 @@ static void vol_store_gain(struct vol_data *cd, const int channels_count)
 
 #if CONFIG_FORMAT_S24LE
 /**
- * \brief HiFi3 enabled volume processing from 24/32 bit to 24/32 or 32 bit.
+ * \brief HiFi4 enabled volume processing from 24/32 bit to 24/32 or 32 bit.
  * \param[in,out] dev Volume base component device.
  * \param[in,out] sink Destination buffer.
  * \param[in,out] source Input buffer.
  * \param[in] frames Number of frames to process.
- * \param[in] attenuation factor for peakmeter adjustment (unused)
+ * \param[in] attenuation factor for peakmeter adjustment
  */
 static void vol_s24_to_s24_s32(struct processing_module *mod, struct input_stream_buffer *bsource,
 			       struct output_stream_buffer *bsink, uint32_t frames,
@@ -66,8 +59,6 @@ static void vol_s24_to_s24_s32(struct processing_module *mod, struct input_strea
 	ae_f32x2 in_sample = AE_ZERO32();
 	ae_f32x2 out_sample = AE_ZERO32();
 	ae_f32x2 volume = AE_ZERO32();
-	ae_f32x2 *buf;
-	ae_f32x2 *buf_end;
 	int i, n, m;
 	ae_f32x2 *vol;
 	ae_valign inu = AE_ZALIGN64();
@@ -79,6 +70,13 @@ static void vol_s24_to_s24_s32(struct processing_module *mod, struct input_strea
 	const int channels_count = audio_stream_get_channels(sink);
 	const int inc = sizeof(ae_f32x2);
 	int samples = channels_count * frames;
+	ae_f32x2 temp;
+	ae_f32x2 *peakvol = (ae_f32x2 *)cd->peak_vol;
+
+	/* Set peakvol(which stores the peak volume data twice) as circular buffer */
+	memset(peakvol, 0, sizeof(ae_f32) * channels_count * 2);
+	AE_SETCBEGIN1(cd->peak_vol);
+	AE_SETCEND1(cd->peak_vol  + channels_count * 2);
 
 	/** to ensure the adsress is 8-byte aligned and avoid risk of
 	 * error loading of volume gain while the cd->vol would be set
@@ -87,20 +85,18 @@ static void vol_s24_to_s24_s32(struct processing_module *mod, struct input_strea
 	if (cd->copy_gain)
 		vol_store_gain(cd, channels_count);
 
-	buf = (ae_f32x2 *)cd->vol;
-	buf_end = (ae_f32x2 *)(cd->vol + channels_count * 2);
-	vol = buf;
+	vol = (ae_f32x2 *)cd->vol;
 	/* Set buf who stores the volume gain data as circular buffer */
-	AE_SETCBEGIN0(buf);
-	AE_SETCEND0(buf_end);
+	AE_SETCBEGIN0(vol);
+	AE_SETCEND0(cd->vol + channels_count * 2);
 
 	bsource->consumed += VOL_S32_SAMPLES_TO_BYTES(samples);
 	bsink->size += VOL_S32_SAMPLES_TO_BYTES(samples);
 
 	while (samples) {
-		m = audio_stream_samples_without_wrap_s24(source, in);
+		m = audio_stream_samples_without_wrap_s32(source, in);
 		n = MIN(m, samples);
-		m = audio_stream_samples_without_wrap_s24(sink, out);
+		m = audio_stream_samples_without_wrap_s16(sink, out);
 		n = MIN(m, n);
 		inu = AE_LA64_PP(in);
 		/* process two continuous sample data once */
@@ -110,7 +106,10 @@ static void vol_s24_to_s24_s32(struct processing_module *mod, struct input_strea
 
 			/* Load the input sample */
 			AE_LA32X2_IP(in_sample, inu, in);
-
+			/* calculate the peak volume*/
+			AE_L32X2_XC1(temp, peakvol, 0);
+			temp = AE_MAXABS32S(in_sample, temp);
+			AE_S32X2_XC1(temp, peakvol, inc);
 			/* Multiply the input sample */
 #if COMP_VOLUME_Q8_16
 			out_sample = AE_MULFP32X2RS(AE_SLAI32S(volume, 7), AE_SLAI32(in_sample, 8));
@@ -132,17 +131,23 @@ static void vol_s24_to_s24_s32(struct processing_module *mod, struct input_strea
 		in = audio_stream_wrap(source, in);
 		out = audio_stream_wrap(sink, out);
 	}
+	for (i = 0; i < channels_count; i++)
+		cd->peak_regs.peak_meter[i] = MAX(cd->peak_vol[i],
+						  cd->peak_vol[i + channels_count])
+						  << (attenuation + PEAK_24S_32C_ADJUST);
+	/* update peak vol */
+	peak_vol_update(cd);
 }
 #endif /* CONFIG_FORMAT_S24LE */
 
 #if CONFIG_FORMAT_S32LE
 /**
- * \brief HiFi3 enabled volume processing from 32 bit to 24/32 or 32 bit.
+ * \brief HiFi4 enabled volume processing from 32 bit to 24/32 or 32 bit.
  * \param[in,out] mod Pointer to struct processing_module
  * \param[in,out] sink Destination buffer.
  * \param[in,out] source Input buffer.
  * \param[in] frames Number of frames to process.
- * \param[in] attenuation factor for peakmeter adjustment (unused)
+ * \param[in] attenuation factor for peakmeter adjustment
  */
 static void vol_s32_to_s24_s32(struct processing_module *mod, struct input_stream_buffer *bsource,
 			       struct output_stream_buffer *bsink, uint32_t frames,
@@ -169,6 +174,13 @@ static void vol_s32_to_s24_s32(struct processing_module *mod, struct input_strea
 						     + bsource->consumed);
 	ae_f32x2 *out = (ae_f32x2 *)audio_stream_wrap(sink, (char *)audio_stream_get_wptr(sink)
 						      + bsink->size);
+	ae_f32x2 temp;
+	ae_f32x2 *peakvol = (ae_f32x2 *)cd->peak_vol;
+
+	/* Set peakvol(which stores the peak volume data twice) as circular buffer */
+	memset(peakvol, 0, sizeof(ae_f32) * channels_count * 2);
+	AE_SETCBEGIN1(cd->peak_vol);
+	AE_SETCEND1(cd->peak_vol  + channels_count * 2);
 
 	/** to ensure the address is 8-byte aligned and avoid risk of
 	 * error loading of volume gain while the cd->vol would be set
@@ -200,7 +212,10 @@ static void vol_s32_to_s24_s32(struct processing_module *mod, struct input_strea
 
 			/* Load the input sample */
 			AE_LA32X2_IP(in_sample, inu, in);
-
+			/* calculate the peak volume*/
+			AE_L32X2_XC1(temp, peakvol, 0);
+			temp = AE_MAXABS32S(in_sample, temp);
+			AE_S32X2_XC1(temp, peakvol, inc);
 #if COMP_VOLUME_Q8_16
 			/* Q8.16 x Q1.31 << 1 -> Q9.48 */
 			mult0 = AE_MULF32S_HH(volume, in_sample);
@@ -225,17 +240,23 @@ static void vol_s32_to_s24_s32(struct processing_module *mod, struct input_strea
 		in = audio_stream_wrap(source, in);
 		out = audio_stream_wrap(sink, out);
 	}
+	for (i = 0; i < channels_count; i++)
+		cd->peak_regs.peak_meter[i] = MAX(cd->peak_vol[i],
+						  cd->peak_vol[i + channels_count]) << attenuation;
+
+	/* update peak vol */
+	peak_vol_update(cd);
 }
 #endif /* CONFIG_FORMAT_S32LE */
 
 #if CONFIG_FORMAT_S16LE
 /**
- * \brief HiFi3 enabled volume processing from 16 bit to 16 bit.
+ * \brief HiFi4 enabled volume processing from 16 bit to 16 bit.
  * \param[in,out] dev Volume base component device.
  * \param[in,out] sink Destination buffer.
  * \param[in,out] source Input buffer.
  * \param[in] frames Number of frames to process.
- * \param[in] attenuation factor for peakmeter adjustment (unused)
+ * \param[in] attenuation factor for peakmeter adjustment (unused for 16bit)
  */
 static void vol_s16_to_s16(struct processing_module *mod, struct input_stream_buffer *bsource,
 			   struct output_stream_buffer *bsink, uint32_t frames,
@@ -263,6 +284,13 @@ static void vol_s16_to_s16(struct processing_module *mod, struct input_stream_bu
 	const int channels_count = audio_stream_get_channels(sink);
 	const int inc = sizeof(ae_f32x2);
 	int samples = channels_count * frames;
+	ae_f32x2 temp;
+	ae_f32x2 *peakvol = (ae_f32x2 *)cd->peak_vol;
+
+	/* Set peakvol(which stores the peak volume data 4 times) as circular buffer */
+	memset(peakvol, 0, sizeof(ae_f32) * channels_count * 4);
+	AE_SETCBEGIN1(cd->peak_vol);
+	AE_SETCEND1(cd->peak_vol  + channels_count * 4);
 
 	/** to ensure the adsress is 8-byte aligned and avoid risk of
 	 * error loading of volume gain while the cd->vol would be set
@@ -303,7 +331,14 @@ static void vol_s16_to_s16(struct processing_module *mod, struct input_stream_bu
 #endif
 			/* Load the input sample */
 			AE_LA16X4_IP(in_sample, inu, in);
-
+			/* calculate the peak volume*/
+			AE_L32X2_XC1(temp, peakvol, 0);
+			temp = AE_MAXABS32S(AE_SEXT32X2D16_32(in_sample), temp);
+			AE_S32X2_XC1(temp, peakvol, inc);
+			/* calculate the peak volume*/
+			AE_L32X2_XC1(temp, peakvol, 0);
+			temp = AE_MAXABS32S(AE_SEXT32X2D16_10(in_sample), temp);
+			AE_S32X2_XC1(temp, peakvol, inc);
 			/* Multiply the input sample */
 			out_sample0 = AE_MULFP32X16X2RS_H(volume0, in_sample);
 			out_sample1 = AE_MULFP32X16X2RS_L(volume1, in_sample);
@@ -314,18 +349,26 @@ static void vol_s16_to_s16(struct processing_module *mod, struct input_stream_bu
 
 			/* store the output */
 			out_sample = AE_ROUND16X4F32SSYM(out_sample0, out_sample1);
-			// AE_SA16X4_IC(out_sample, outu, out);
 			AE_SA16X4_IP(out_sample, outu, out);
 		}
 		AE_SA64POS_FP(outu, out);
 		samples -= n;
-		bsource->consumed += VOL_S16_SAMPLES_TO_BYTES(n);
-		bsink->size += VOL_S16_SAMPLES_TO_BYTES(n);
 		in = audio_stream_wrap(source, in);
 		out = audio_stream_wrap(sink, out);
+		bsource->consumed += VOL_S16_SAMPLES_TO_BYTES(n);
+		bsink->size += VOL_S16_SAMPLES_TO_BYTES(n);
 	}
+	for (i = 0; i < channels_count; i++) {
+		m = MAX(cd->peak_vol[i], cd->peak_vol[i + channels_count]);
+		m = MAX(m, cd->peak_vol[i + channels_count * 2]);
+		m = MAX(m, cd->peak_vol[i + channels_count * 3]);
+		cd->peak_regs.peak_meter[i] = m;
+	}
+	/* update peak vol */
+	peak_vol_update(cd);
 }
 #endif /* CONFIG_FORMAT_S16LE */
+
 const struct comp_func_map volume_func_map[] = {
 #if CONFIG_FORMAT_S16LE
 	{ SOF_IPC_FRAME_S16_LE, vol_s16_to_s16 },
